requirements.txt
google-api-python-client==2.105.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
html2text==2020.1.16
openai==1.12.0
python-dotenv==0.21.0

# You can install these by running:
# $ pip3 install -r requirements.txt
